SCRUM-89 Consolidating code base
Consolidated code base with a focus on making the coding style and naming convention more consistent.
Renamed Autobot folder to ReactApp.
Changed jsx files in the pages folder to be in PascalCase and end with the suffix "Page".
Moved SignUpPage.jsx from its own folder into the pages folder.
Moved SignUpPage.css into the styles folder.
Changed tab spacing to two for all jsx files.
Cleaned up code by removing whitespaces, alphabetizing/cleaning imports, and wrapping long lines for better readability.
Added PasswordTextField.jsx in the components folder for better reusability for LoginPage.jsx and SignUpPage.jsx.

Moved and modified some components to improve UX flow.
Added some fields into the table displayed in ActivityLogPage.jsx
Moved RobotSettingPage.jsx setting icon button from ActivityLogPage.jsx to ControlPage.jsx.
Moved FPV/Control button from ActivityLogPage.jsx to the individual buttons below each displayed robot on HomePage.jsx.
Added an activity log button to HomePage.jsx that allows users to navigate to ActivityLogPage.jsx.

It could be a good idea to add a sprocket icon button to each displayed robot on HomePage.jsx in the future.
There is definitely more room for further consolidation. E.g. Adding more component templates and ensuring that MUI is being utilized throughout the code base.
