+++ ReactApp/src/components/AddButton.jsx
@@ -1,16 +1,17 @@
 import React from "react";
+import { useNavigate } from "react-router-dom";
+
 import IconButton from "@mui/material/IconButton";
 import AddCircleOutlineOutlinedIcon from '@mui/icons-material/AddCircleOutlineOutlined';
-import { useNavigate } from "react-router-dom";
 
 {/* 
   AddButton is a template for a plus surrounded by a circle icon button. This will
-  typically be used to send users to an add page. AddButton takes a string argument
+  typically be used to send users to an add page. AddButton passes path as a prop
   that represents the path of the page the button will take the user to.
 */}
 const AddButton = ({path="/"}) => { //defaults to "/" page if no url specified
   const navigate = useNavigate();
-    
+
   return (
     <IconButton
       onClick={() => navigate(path)}
@@ -24,6 +25,6 @@ const AddButton = ({path="/"}) => { //defaults to "/" page if no url specified
       <AddCircleOutlineOutlinedIcon fontSize="large"/>
     </IconButton>
   );
-}
+};
 
 export default AddButton;
