+++ ReactApp/src/components/DisplayCard.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Card, CardMedia, CardContent, Typography, Box } from "@mui/material";
+import { Box, Card, CardContent, CardMedia, Typography } from "@mui/material";
 
 /* 
  Card component to display the detected object and its metadata.
