+++ ReactApp/src/components/SettingsButton.jsx
@@ -1,16 +1,17 @@
 import React from "react";
-import IconButton from "@mui/material/IconButton";
-import SettingsOutlinedIcon from '@mui/icons-material/SettingsOutlined';
 import { useNavigate } from "react-router-dom";
 
+import IconButton from "@mui/material/IconButton";
+import SettingsOutlinedIcon from '@mui/icons-material/SettingsOutlined';
+
 {/* 
-  Template for a sprocket icon button. This will typically be used to send
-  users to a settings page. SettingsButton takes a string argument that
-  represents the path of the page the button will take the user to.
+  SettingsButton is a template for a sprocket icon button. This will typically
+  be used to send users to a settings page. SettingsButton passes path as a prop
+  that represents the path of the page the button will take the user to.
 */}
 const SettingsButton = ({path="/"}) => { //defaults to "/" page if no url specified
   const navigate = useNavigate();
-    
+
   return (
     <IconButton
       onClick={() => navigate(path)}
