+++ ReactApp/src/main.jsx
@@ -1,13 +1,13 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { BrowserRouter as Router } from "react-router-dom"; //probably need this
-import './index.css'
+import { BrowserRouter as Router } from "react-router-dom";
 import App from './App.jsx'
+import './index.css'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
     <Router>
       <App />
     </Router>
-  </StrictMode>,
+  </StrictMode>
 )
