+++ ReactApp/src/pages/RobotSettingPage.jsx
@@ -1,9 +1,11 @@
 import React from "react";
-import { Box, Typography, Button, IconButton } from "@mui/material";
-import ArrowBackIcon from "@mui/icons-material/ArrowBack";
 import { useNavigate } from "react-router-dom";
 
-function RobotSettings() {
+import { Box, Typography, Button, IconButton } from "@mui/material";
+import BackButton from "../components/BackButton";
+
+
+function RobotSettingPage() {
   const navigate = useNavigate();
 
   return (
@@ -11,26 +13,13 @@ function RobotSettings() {
       sx={{
         width: "100%",
         minHeight: "100vh",
-        bgcolor: "#FFFFFF",
         display: "flex",
         flexDirection: "column",
         alignItems: "center",
-        position: "relative",
-        paddingTop: 4,
+        paddingTop: 7
       }}
     >
-      <IconButton
-        onClick={() => navigate("/ActivityLog")}
-        sx={{
-          position: "absolute",
-          top: 0,
-          right: -500,
-          color: "black",
-          padding: "40",
-        }}
-      >
-        <ArrowBackIcon fontSize="large" />
-      </IconButton>
+      <BackButton path="/ActivityLogPage" />
 
       <Button
         // Delete Robot Button
@@ -45,7 +34,7 @@ function RobotSettings() {
           mb: 2,
           "&:hover": {
             backgroundColor: "#C00F0C",
-          },
+          }
         }}
       >
         Delete Robot
@@ -63,7 +52,7 @@ function RobotSettings() {
           mb: 2,
           "&:hover": {
             backgroundColor: "#C00F0C",
-          },
+          }
         }}
       >
         Shutdown
@@ -81,7 +70,7 @@ function RobotSettings() {
           mb: 2,
           "&:hover": {
             backgroundColor: "#4A4A4A",
-          },
+          }
         }}
       >
         Update Software
@@ -98,8 +87,8 @@ function RobotSettings() {
           height: 69,
           mb: 2,
           "&:hover": {
-            backgroundColor: "#4A4A4A",
-          },
+            backgroundColor: "#4A4A4A"
+          }
         }}
       >
         Change Network Name
@@ -109,7 +98,7 @@ function RobotSettings() {
         sx={{
           mt: 25,
           mb: 4,
-          textAlign: "center",
+          textAlign: "center"
         }}
       >
         <Typography
@@ -117,27 +106,29 @@ function RobotSettings() {
             color: "#000000",
             fontFamily: "Inter",
             fontSize: 16,
-            lineHeight: "100%",
+            lineHeight: "100%"
           }}
         >
           Operating System: Ubuntu 24.04
         </Typography>
+
         <Typography
           sx={{
             color: "#000000",
             fontFamily: "Inter",
             fontSize: 16,
-            lineHeight: "100%",
+            lineHeight: "100%"
           }}
         >
           ROS Version: 19.02 Jazzy Jalisco
         </Typography>
+
         <Typography
           sx={{
             color: "#000000",
             fontFamily: "Inter",
             fontSize: 16,
-            lineHeight: "100%",
+            lineHeight: "100%"
           }}
         >
           Date Added To Network: 02/27/2025
@@ -145,6 +136,6 @@ function RobotSettings() {
       </Box>
     </Box>
   );
-}
+};
 
-export default RobotSettings;
+export default RobotSettingPage;
