+++ ReactApp/src/styles/RobotCard.css
