+++ ReactApp/vite.config.js
